refactor(ui): rename maintenance view to tsx
Issue #2403 expects this last h()-based view to match the repo's\nJSX convention, so the file now uses JSX and the .tsx extension.\n\nRefs #2403\n\nCo-authored-by: Copilot <223556219+Copilot@users.noreply.github.com>
